+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Added `fonts_as_import()` to create stylesheet urls for embedding of fonts in
   HTML and SVG
+* Added two C-level functions for getting glyph outline and bitmap information
 
 # systemfonts 1.2.2